deadLine类型改为long

+++ base/src/main/java/com/rt/base/bean/ParkingLotResultBean.kt
@@ -8,7 +8,7 @@ data class ParkingLotBean(
     val carColor: String,
     val carLicense: String,
     val cleared: String,
-    val deadLine: Int,
+    val deadLine: Long,
     val orderNo: String,
     val parkingNo: String,
     val state: String
